+++ src/dynd/dtypes/base_string_dtype.cpp
@@ -21,7 +21,7 @@ std::string base_string_dtype::get_utf8_string(const char *metadata, const char 
     return string_range_as_utf8_string(get_encoding(), begin, end, errmode);
 }
 
-size_t base_string_dtype::get_iterdata_size(int DYND_UNUSED(ndim)) const
+size_t base_string_dtype::get_iterdata_size(size_t DYND_UNUSED(ndim)) const
 {
     return 0;
 }
